Tetris Game using Linux
Environment : Ubuntu 16.04.6 LTS (GNU/Linux 4.4.0-150-generic x86_64)
Library : ncurses.h, unistd.h

This toy project is to play a tetris game. It is developed for the project of the class - CSE3013 (컴퓨터공학실험I) in Sogang University.
You have to place those files in a same folder, and command "make", which produces a binary file - a.out.
Finally, just run the binary file and enjoy it!